5547 - Make links redux compliant with new store conventions
